Renamed Table Content -> Body folder.
