+++ tools/gulptasks/dashboards/scripts-dts/datagrid.src.d.ts
@@ -16,8 +16,8 @@ import './es-modules/Grid/Pro/Credits/CreditsProComposition';
 export { /** @deprecated Use `Grid` instead. */ default as DataGrid } from './es-modules/Grid/Core/Grid.js';
 export { default as Grid } from './es-modules/Grid/Core/Grid.js';
 export { default as Column } from './es-modules/Grid/Core/Table/Column.js';
-export { default as TableRow } from './es-modules/Grid/Core/Table/Content/TableRow.js';
-export { default as TableCell } from './es-modules/Grid/Core/Table/Content/TableCell.js';
+export { default as TableRow } from './es-modules/Grid/Core/Table/Body/TableRow.js';
+export { default as TableCell } from './es-modules/Grid/Core/Table/Body/TableCell.js';
 export { default as Options } from './es-modules/Grid/Core/Options.js';
 
 export { default as DataConnector } from './es-modules/Data/Connectors/DataConnector';
