+++ tools/gulptasks/scripts-dts/grid-lite.src.d.ts
@@ -10,8 +10,8 @@ import Defaults from "./es-modules/Grid/Core/Defaults";
 
 export { default as Grid } from './es-modules/Grid/Core/Grid.js';
 export { default as Column } from './es-modules/Grid/Core/Table/Column.js';
-export { default as TableRow } from './es-modules/Grid/Core/Table/Content/TableRow.js';
-export { default as TableCell } from './es-modules/Grid/Core/Table/Content/TableCell.js';
+export { default as TableRow } from './es-modules/Grid/Core/Table/Body/TableRow.js';
+export { default as TableCell } from './es-modules/Grid/Core/Table/Body/TableCell.js';
 export { default as Options } from './es-modules/Grid/Core/Options.js';
 
 export { default as DataConnector } from "./es-modules/Data/Connectors/DataConnector";
