+++ ts/Dashboards/datagrid.d.ts
@@ -1,6 +1,6 @@
 export { /** @deprecated Use `Grid` instead */ default as DataGrid } from '../Grid/Core/Grid.js';
 export { default as Grid } from '../Grid/Core/Grid.js';
 export { default as Column } from '../Grid/Core/Table/Column.js';
-export { default as TableRow } from '../Grid/Core/Table/Content/TableRow.js';
-export { default as TableCell } from '../Grid/Core/Table/Content/TableCell.js';
+export { default as TableRow } from '../Grid/Core/Table/Body/TableRow.js';
+export { default as TableCell } from '../Grid/Core/Table/Body/TableCell.js';
 export { default as Options } from '../Grid/Core/Options.js';
