+++ ts/Grid/Core/Table/Cell.ts
@@ -23,7 +23,7 @@
  * */
 
 import type DataTable from '../../../Data/DataTable';
-import type TableRow from './Content/TableRow';
+import type TableRow from './Body/TableRow';
 
 import Column from './Column';
 import Row from './Row';
